### 2021-08-23 #3648 @whedon
[PRE REVIEW]: PDLSM-FEM: Solver of Coupled Peridynamics Least Squares Minimization with Finite Element Method
**Submitting author:** @QibangLiu  (<a href="http://orcid.org/0000-0001-7935-7907">Qibang Liu</a>)
**Repository:** <a href="https://github.com/QibangLiu/PDLSM-FEM_mpi" target ="_blank">https://github.com/QibangLiu/PDLSM-FEM_mpi</a>
**Version:** v1.0

Thanks for submitting your paper to JOSS @QibangLiu . **Currently, there isn't an JOSS editor assigned** to your paper.
