### 2021-08-23 #3648 @danielskatz
👋 @QibangLiu - Please note that your paper does not compile.  Your header has a lot of blank lines that may be causing problems.  Please follow the example in https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

In addition, you could work on the possibly missing DOI that whedon suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again

When you have made changes to the paper, use command `@whedon generate pdf` to make a new PDF.

Whedon commands need to be the first entry in a new comment.
