+++ JOSS_paper/paper.bib
@@ -1,164 +1,8 @@
-@article{Goyal_2017,
-  title={Design and Control of Growth Adaptable Artificial Gravity Space Habitat},
-  author={Goyal, R. and Bryant, T. and Majji, M. and Skelton, R. and Longman, A.},
-  booktitle={AIAA SPACE and Astronautics Forum and Exposition},
-  pages={5141},
-  year={2017},
-  doi ={10.2514/6.2017-5141}, 
-}
-
-@article{Chen_2018_Habitat,
-  title={Energy Analysis of Growth Adaptable Artificial Gravity Space Habitat},
-  author={Chen, Muhao and Shen, Yuling and Goyal, Raman and Majji, Manoranjan and Skelton, Robert E.},
-  booktitle={AIAA SPACE and Astronautics Forum and Exposition},
-  pages={5109},
-  year={2018},
-  doi ={10.2514/6.2018-5109},
-}
-
-@article{Goyal_2018_ASCE,
-title={Dynamics of Class 1 Tensegrity Systems Including Cable Mass},
-author={Goyal, R. and Skelton, R. E.},
-booktitle={16th Biennial ASCE Conference - Earth and Space 2018},
-pages = {868 - 876},
-DOI = {10.1061/9780784481899.082},
-URL = {https://ascelibrary.org/doi/abs/10.1061/9780784481899.082},
-year={2018}
-}
-
-@article{Karnan_2017, 
-author={Karnan, H. and Goyal, R. and Majji, M. and Skelton, R. E. and Singla, P.}, 
-journal={2017 IEEE/RSJ-IROS}, 
-title={Visual feedback control of tensegrity robotic systems}, 
-year={2017}, 
-pages={2048-2053}, 
-doi={10.1109/IROS.2017.8206022}, 
-ISSN={2153-0866}, 
-month={Sept},
-}
-
-@article{Goyal_2019_Buckling,
-  title={Analytical Study of Tensegrity Lattices for Mass-Efficient Mechanical Energy Absorption},
-  author={Goyal, R. and Hernandez, E. A. P. and Skelton, R. E.},
-  journal={International Journal of Space Structures},
-  pages={1-19},
-  DOI = {10.1177/0956059919845330},
-  year={2019}
-}
-
-@INPROCEEDINGS{Sabelhaus_2017, 
-author={Sabelhaus, A. P. and Akella, A. K. and Ahmad, Z. A. and SunSpiral, V.}, 
-booktitle={2017 American Control Conference (ACC)}, 
-title={Model-Predictive Control of a flexible spine robot}, 
-year={2017}, 
-volume={}, 
-number={}, 
-pages={5051-5057}, 
-doi={10.23919/ACC.2017.7963738}, 
-ISSN={2378-5861}, 
-month={May},}
-
-@article{Goyal_Dynamics_2019,
-  title={Tensegrity System Dynamics with Rigid Bars and Massive Strings},
-  author={Goyal, R. and Skelton, R. E.},
-  journal={Multibody System Dynamics},
-  volume={46},
-  number={3},
-  pages={203-228},
-  DOI = {10.1007/s11044-019-09666-4},
-  year={2019}
-}
-
-@book{Skelton_2009_Tensegrity_Book,
-  title={Tensegrity Systems},
-  author={Skelton, R. E. and de Oliveira, M. C.},
-  ISBN = {9780387742427},
-  year={2009},
-  publisher={Springer US}
-}
-
-@article{Vytas_2013,
-  title={Tensegrity based probes for planetary exploration: Entry, descent and landing ({EDL}) and surface mobility analysis},
[…]
