### 2021-08-23 cef6dcd
Revert "Update paper.md"
This reverts commit 706afa3330ac57c076921971464996f69794d15b.
