+++ README.md
@@ -1,0 +1,6 @@
+# PDLSM-FEM_mpi
+PDLSM-FEM solver is an open-source parallel implementation of coupled peridynamics least squares minimization and finite element method (PDLSM-FEM) 
+in 2D and 3D by MPI technique. This solver is written in a C++ environment on cross platforms (Windows, Linux). It includes implicit static, explicit
+dynamic, and implicit dynamic solvers for structure analysis under displacement or traction loading. 
+
+See "PDLSM_FEM_Solver manual.pdf" for more details.
