### 2021-08-24 #3648 @prashjha
Hey @karthikncsu, would you be interested in reviewing this submission for JOSS (Journal of Open Source Software)? In JOSS, reviews are open and transparent. As a reviewer, you have to assess the submission using a checklist. You can find more information about the reviewing process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html# If you need any help, let me know.
