### 2021-08-24 #3648 @danielskatz
Hi @prashjha - I'm a bit confused by what's going on here. Normally in JOSS, the editor finds at least 2 reviewers, but is not one of the reviewers.
