### 2021-08-24 #3648 @danielskatz
We normally only have the editor act as a reviewer if we lose a reviewer during the process and need someone to step in.  Ideally, the two reviewers and the editor are separate people to have two independent points of view and a neutral editor to make the process work. Unless there's a reason you can't find 2 reviewers, you should.
