### 2021-08-29 #3668 @whedon
[REVIEW]: PDLSM-FEM: Solver of Coupled Peridynamics Least Squares Minimization with Finite Element Method
**Submitting author:** @QibangLiu  (<a href="http://orcid.org/0000-0001-7935-7907">Qibang Liu</a>)
**Repository:** <a href="https://github.com/QibangLiu/PDLSM-FEM_mpi" target ="_blank">https://github.com/QibangLiu/PDLSM-FEM_mpi</a>
**Version:** v1.0

- [x] **Contribution and authorship:** Has the submitting author (@QibangLiu ) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Contribution and authorship:** Has the submitting author (@QibangLiu ) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] #3765

- [x] **Contribution and authorship:** Has the submitting author (@QibangLiu ) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] #3810
