+++ SourceFiles/Solver/fioFile.cpp
@@ -122,7 +122,7 @@ void fioFiles::excuteCMD(datModel& o_dat, vector<string>& tokens)
 			}
 			else if (tokens[0] == "SETSOLVING") //setsolver cmd
 			{
-				if (tokens.size() < 6)
+				if (tokens.size() < 7)
 				{
 					if (ci_rank == 0)
 					{
@@ -132,10 +132,15 @@ void fioFiles::excuteCMD(datModel& o_dat, vector<string>& tokens)
 				else
 				{
 					o_dat.cd_dt = atof(tokens[1].c_str());
+					if (o_dat.ci_solvFlag==2)
+					{
+						o_dat.cd_dt = 1;
+					}
 					o_dat.ci_numTstep= atoi(tokens[2].c_str());
 					o_dat.ci_savefrequence= atoi(tokens[3].c_str());
 					o_dat.op_getGeomP()->cd_factor= atof(tokens[4].c_str());
 					o_dat.cd_NLF = atof(tokens[5].c_str());
+					o_dat.cd_me= atof(tokens[6].c_str());
 				}
 			}
 			else if (tokens[0] == "NEWMARK") //setsolver cmd
@@ -239,6 +244,7 @@ void fioFiles::excuteCMD(datModel& o_dat, vector<string>& tokens)
 						velo = atof(tokens[2].c_str());
 						o_dat.op_getEssenBC(ID)->cb_varing = true;
 						o_dat.op_getEssenBC(ID)->cd_velocity = velo;
+						o_dat.op_getEssenBC(ID)->cd_maxV = velo;
 					}
 					else if (ci_rank==0)
 					{
@@ -491,18 +497,18 @@ void fioFiles::writeResults(datModel &o_dat)
 		writeReslutsTOTAL_vtk(o_dat, "static");
 		writeResultsPD_vtk(o_dat);
 		writeResultsFEM_vtk(o_dat);
-		writeUofNode(o_dat);
-		writeSigofNode(o_dat);
+		writeUofNode(o_dat, "static");
+		writeSigofNode(o_dat, "static");
 	}
 	
 
 }
 
 
-void fioFiles::writeUofNode(datModel &o_dat)
+void fioFiles::writeUofNode(datModel &o_dat, string nTstep)
 {
 	char fileNAME[_MAX_PATH];
-	sprintf(fileNAME, "%s.disp", cc_fileNAME);
+	sprintf(fileNAME, "%s_%s.disp", cc_fileNAME, nTstep.c_str());
 	ofstream fout(fileNAME);
 	fout << "*****The final displacement of  point*****" << endl;
 	fout << std::left;
@@ -541,10 +547,10 @@ bool fioFiles::is_little_endian()
 	return (c.b == 1);
 }
 
-void fioFiles::writeSigofNode(datModel &o_dat)
+void fioFiles::writeSigofNode(datModel &o_dat, string nTstep)
 {
 	char fileNAME[_MAX_PATH];
-	sprintf(fileNAME,"%s.sig", cc_fileNAME);
+	sprintf(fileNAME,"%s_%s.sig", cc_fileNAME, nTstep.c_str());
 	ofstream fout(fileNAME);
 	fout << "*****The stress of  point*****" << endl;
 	fout << std::left;
@@ -566,6 +572,7 @@ void fioFiles::writeReslutsTOTAL_vtk(datModel &o_dat, string nTstep)
 	if (o_dat.cb_vtkBinary)
 	{
 		writeReslutsTOTAL_vtk_Binary(o_dat, nTstep);
+		writeSigofNode(o_dat, nTstep);
 	}
 	else
 	{
@@ -656,7 +663,7 @@ void fioFiles::writeReslutsTOTAL_vtk_Binary(datModel& o_dat, string nTstep)
[…]
