### 2021-09-26 #3668 @karthikncsu
@QibangLiu

Installation of the PDLSM-FEM solver software requires setting up the paths for MKLROOT and MPIROOT as mentioned in the manual: page 7 section 2.2 Build on Linux

export MKLROOT =/ homes / MKL_2020 / compilers_and_libraries_2020 .1.217/ linux / mkl
export MPIROOT =/ opt / software / OpenMPI /4.0.5 - iccifort -2020.4.304

I suppose these pathnames are available in the older version of intel parallel studio.

Intel had discontinued the old intel parallel studio product and have released the open-source oneAPI tool kit, replacing Intel Parallel Studio XE on December 8, 2020.

https://software.intel.com/content/www/us/en/develop/tools/oneapi/all-toolkits.html
https://en.wikipedia.org/wiki/Intel_Parallel_Studio

I faced similar issues recently when my intel parallel studio license has expired and had to change my code to be compatible with oneAPI. I am not able to install the package as I don't have the intel parallel studio license. Please modify the code and the documentation with the new paths for the mkl and iccifort filenames in the open API installation directory.

Please let me know if there is an alternate way of installation of the software.
