### 2021-09-26 e4992b8
update Find MPI.cmake and the manual file
