### 2021-09-26 #3668 @QibangLiu
@karthikncsu  Thanks for your comments. I didn't realize they closed the old version and released the oneAPI toolkit.
The principles to set environment variables are that:
 (a) Under the path "MKLROOT/include", there is the "mkl.h file", and under the path "MKLROOT/lib/intel64", there are libraries as the MKL link line advisor show, such as "libmkl_intel_ilp64.so", "libmkl_core.so".
(b) Under the path "MPIROOT/include", there is the header file "mpi.h",  and under the path "MPIROOT/lib", there is the library "libmpi.so".
This explanation is added in the manual file.

For the MPIROOT, the OpenMPI is installed under the path "/opt / software / OpenMPI /4.0.5 - iccifort -2020.4.304" on my machine, but not installed with the intel parallel studio product.
Anyway, I modified the values based on the oneAPI toolkit setting as your request:
export MKLROOT=/homes/qibangliu/intel/oneapi/mkl/2021.3.0
export MPIROOT=/homes/qibangliu/intel/oneapi/mpi/2021.3.0

I noted that the file "libmpi.so" is under the path /homes/qibangliu/intel/oneapi/mpi/2021.3.0/lib/release, so the
HINTS value in the FindMPI.cmake file should be modified as:
find_library(MPI_LIBRARY NAMES libmpi.so  HINTS $ENV{MPIROOT}/lib $ENV{MPIROOT}/lib/release)

The updates were posted on GitHub.
