+++ JOSS_paper/paper.bib
@@ -118,3 +118,16 @@
 	date = {2021-07},
 	langid = {english},
 }
+
+
+@article{liu_coupled_2021,
+	title = {Coupled Peridynamics Least Square Minimization with Finite Element Method in 3D and Implicit Solutions by Message Passing Interface},
+	rights = {All rights reserved},
+	issn = {2522-896X, 2522-8978},
+	doi = {10.1007/s42102-021-00060-3},
+	journaltitle = {Journal of Peridynamics and Nonlocal Modeling},
+	shortjournal = {J Peridyn Nonlocal Model},
+	author = {Liu, Qibang and Xin, X. J. and Ma, Jeff},
+	date = {2021-09-06},
+	langid = {english},
+}
