+++ HeaderFiles/Solver/pdsolve.h
@@ -26,15 +26,6 @@ struct critStru
 		return sd_value > s_a.sd_value; //min top heap;
 	}
 };
-
-struct GausPoint_Jint
-{
-	//this structurc is for 2D J-integral, represent the path points.
-	double x[3];
-	double theta;
-	double sigma[3];
-	double DG[4];//deformation gradient;
-};
 class pdsolve
 {
 public:
@@ -43,30 +34,17 @@ public:
 	
 	//====data model setting=============================================
 	void setDatModel(datModel& o_dat); // set the data model, must be called before solving;
-	void Setdof_Index(datModel& o_dat);//set each dof's equation position;
-	void setDeltaMaxMin_EleV_NV(datModel& o_dat);// find out the max and min Delta. Got ele volume and  node volume
-	void findDomainDimen(datModel& o_dat);// get the domain dimension size;
-	void setBlock(datModel& o_dat);// allocate memory for block, and put elements and elements into block;
-	void initializePDE_FE(datModel& o_dat);//judge FE and PDE, then put the idx into vector, based on intersection between element and crack;
-	void updatePDBE(datModel& o_dat);//update PDBE and assign the pointers, based on the civ_pdeIDX;
-	void UpdateNodeType(datModel& o_dat);
-	void initialPDN_Fami(datModel& o_dat);//put the PD node idx into civ_pdNodeIDX, set nodes' famyID, set numFami, allocaMemory for familyies, cal PD node volume
-	void updatePDN_Fami(datModel& o_dat);
-	void initialBondState(datModel& o_dat);
-
-	void calVolumeOfNode(datModel& o_dat);// calculate volume of pd node; set horizon size;
+	void setFEID_PDEID(datModel& o_dat);//set the FE id and PDE id for each cores;
+	void findDomainDimen(datModel &o_dat);// get the domain dimension size;
+	void calVolumeOfNode(datModel& o_dat);// calculate volume of pd node;
 	void setPDNODEandnumFami(datModel& o_dat);// set the node type and memory allocate for family;
+	void Setdof_Index(datModel& o_dat);//set each dof's equation position;
+	void setDeltaMaxMin(datModel& o_dat);// find out the max and min Delta;
 	void setBlockAndFami(datModel& o_dat);// initialize block;
-	
-	void setNoFailRegion(datModel& o_dat);// may not need for EFEM-PD;
+	void initialBondState(datModel& o_dat);
+	void setNoFailRegion(datModel& o_dat);
 	//============PD algorithem===============================================;
 	//====some auxiliary functions
-	static bool b_vcom1(int* a, int* b); //must be static or external funtion
-	static bool b_vcom_GP(const GausPoint_Jint &a, const GausPoint_Jint& b); //must be static or external funtion
-	
-	void assignPDBE(datModel& o_dat);
-	void updatePDEbyCrack(double crack[][3], datModel& o_dat);
-	bool b_eleIntersectCrack(int eleID, double crack[][3], datModel& o_dat);
 	long long int findCSRIndexOfMat(int rowIndex, int colIndex);
 	double calArea(double vec1[], double vec2[]);
 	void shapeFunctionQuad4N(double N[], double p, double q);
@@ -100,7 +78,6 @@ public:
 	void assembleInterWorkPD_CSRformat(datModel& o_dat,double *U_N);
 
 	void assemblePDBEwork3D_CSRformat(datModel& o_dat, double* U_N);
-	void normMatArea(Matrix* Nmat, int conNid[], double xN[][3]);
[…]
