+++ HeaderFiles/Solver/pdPDBEs.h
@@ -6,10 +6,9 @@ public:
 	~pdPDBEs();
 	void getNodeID(int nID[]);
 	int i_getNumnode();
-	int* ci_nodeID;
 private:
 	pdPDBEs();
-	
+	int *ci_nodeID;
 	int /*ci_normDire,*/ ci_numNode;
 };
 
