+++ HeaderFiles/Solver/pdFamily.h
@@ -25,10 +25,9 @@ public:
 	vector<int>civ_NID;
 	int* cip_bondState;//bond state ;
 	bool cb_allowFail;
-	double cd_v;
 private: 
 	double cd_delta;
-	int ci_famID;//no need any more;
+	int ci_famID;
 	
 	
 };
