### 2021-09-27 #3668 @karthikncsu
Hi @QibangLiu  Thank you for updating the manual with the settings to run the PDLSM-FEM with one API.

I tried to install on my ubuntu system and it returned below error:

/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `GOMP_loop_end_nowait'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `GOMP_loop_ordered_static_start'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `omp_get_thread_num'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `GOMP_loop_dynamic_next'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `GOMP_sections_next'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `GOMP_parallel_end'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `omp_get_num_threads'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `omp_set_max_active_levels'
/opt/intel/oneapi/mkl/2021.3.0//lib/intel64/libmkl_gnu_thread.so: undefined reference to `GOMP_parallel_sections_start'
collect2: error: ld returned 1 exit status
CMakeFiles/PDLSM-FEM.dir/build.make:854: recipe for target 'PDLSM-FEM' failed
make[2]: *** [PDLSM-FEM] Error 1
CMakeFiles/Makefile2:67: recipe for target 'CMakeFiles/PDLSM-FEM.dir/all' failed
make[1]: *** [CMakeFiles/PDLSM-FEM.dir/all] Error 2
Makefile:83: recipe for target 'all' failed
make: *** [all] Error 2

Let me know a way around to resolve this. I would also recommend you to provide requirements files in the Github repo.
