### 2021-09-27 #3668 @TLCFEM
need flag -fopenmp
