+++ cmake/FindMPI.cmake
@@ -15,9 +15,9 @@ endif()
 if(WIN32)
    find_library(MPI_LIBRARY NAMES msmpi.lib  HINTS $ENV{MPIROOT}/Lib/x64)
 elseif(UNIX)
-   find_library(MPI_LIBRARY NAMES libmpi.so  HINTS $ENV{MPIROOT}/lib $ENV{MPIROOT}/lib/release)
+   find_library(MPI_LIBRARY NAMES libmpi.so  HINTS $ENV{MPIROOT}/lib/release $ENV{MPIROOT}/lib)
 endif()
-
+#message(STAUS "==$ENV{MPIROOT}")
 set(MPI_LIBRARIES ${MPI_LIBRARY})
 set(MPI_INCLUDE_DIRS ${MPI_INCLUDE_DIR})
 include(FindPackageHandleStandardArgs)
