+++ cmake/CMakeLists.txt
@@ -46,6 +46,6 @@ ${MPI_LIBRARIES}
 if(WIN32)   
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /DMKL_ILP64")
 elseif(UNIX)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -lgomp -lpthread -lm -ldl -DMKL_ILP64 -m64")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -lpthread -lm -ldl -DMKL_ILP64 -m64")
 endif()
 SET(CMAKE_BUILD_TYPE "Release")
