### 2021-09-28 #3668 @QibangLiu
@karthikncsu @TLCFEM The error is because the library "libmkl_gnu_thread.so" requires libgomp. I have tried to use combination of OpenMP and MPI to parallel the code. It is my bad that I forgot to remove the requirement of libmkl_gnu_thread.

I have replaced the libmkl_gnu_thread.so by libmkl_sequential.so.

I tried to extract the required files, but some of them have very large sizes, that are difficult to post on Github.

By the way, if your use the MPI version in the OneAPI package, you may have to replace the requirement library "libmkl_blacs_openmpi_ilp64.so" by "libmkl_blacs_intelmpi_ilp64.so", following the instruction in the manual file (the head of section 2), as the MPI in the OneAPI package is "Intel MPI", but the file "FindMKL.cmake" set it for OpenMPI.

Updates are posted on Github.
