### 2021-10-10 e9d69f8
Add features for macOS, and a example for paper
