+++ JOSS_paper/paper.bib
@@ -131,3 +131,17 @@
 	date = {2021-09-06},
 	langid = {english},
 }
+
+@article{ayatollahi_analysis_2009,
+	title = {Analysis of a new specimen for mixed mode fracture tests on brittle materials},
+	volume = {76},
+	issn = {00137944},
+	doi = {10.1016/j.engfracmech.2009.02.016},
+	pages = {1563--1573},
+	number = {11},
+	journaltitle = {Engineering Fracture Mechanics},
+	shortjournal = {Engineering Fracture Mechanics},
+	author = {Ayatollahi, M.R. and Aliha, M.R.M.},
+	date = {2009-07},
+	langid = {english},
+}
