+++ README.md
@@ -1,6 +1,6 @@
 # PDLSM-FEM_mpi
 [PDLSM-FEM solver](https://github.com/QibangLiu/PDLSM-FEM_mpi) is an open-source parallel implementation of coupled peridynamics least squares minimization and finite element method (PDLSM-FEM) 
-in 2D and 3D by MPI technique. This solver is written in a C++ environment on cross platforms (Windows, Linux). It includes implicit static, explicit
+in 2D and 3D by MPI technique. This solver is written in a C++ environment on cross platforms. It includes implicit static, explicit
 dynamic, and implicit dynamic solvers for structure analysis under displacement or traction loading. 
 
   - `Pre-process` - It requires a mesh data file with the specified format. Users may use commercial finite element codes such as ANSYS, ABAQUS to generate the mesh data file.  
