+++ cmake/CMakeLists.txt
@@ -45,7 +45,9 @@ ${MPI_LIBRARIES}
 #===set compiler options=====================
 if(WIN32)   
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /DMKL_ILP64")
-elseif(UNIX)
+elseif(UNIX AND NOT APPLE)
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -lpthread -lm -ldl -DMKL_ILP64 -m64")
+elseif(APPLE)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -lpthread -lm -ldl -DMKL_ILP64 -m64")
 endif()
 SET(CMAKE_BUILD_TYPE "Release")
