+++ cmake/FindMPI.cmake
@@ -6,16 +6,20 @@
 #  MPI_FOUND        - True if mpi found.
 #
 
+#==========find includ dir===========
 if(WIN32)
-   find_path(MPI_INCLUDE_DIR mpi.h HINTS $ENV{MPIROOT}/Include)
+   find_path(MPI_INCLUDE_DIR mpi.h HINTS $ENV{MPIROOT}/Include $ENV{MPIROOT}/include)
 elseif(UNIX)
    find_path(MPI_INCLUDE_DIR mpi.h HINTS $ENV{MPIROOT}/include)
 endif()
 
+#============find libraries
 if(WIN32)
    find_library(MPI_LIBRARY NAMES msmpi.lib  HINTS $ENV{MPIROOT}/Lib/x64)
-elseif(UNIX)
+elseif(UNIX AND NOT APPLE)
    find_library(MPI_LIBRARY NAMES libmpi.so  HINTS $ENV{MPIROOT}/lib/release $ENV{MPIROOT}/lib)
+elseif(APPLE)
+   find_library(MPI_LIBRARY NAMES libmpi.dylib  HINTS $ENV{MPIROOT}/lib/release $ENV{MPIROOT}/lib)
 endif()
 #message(STAUS "==$ENV{MPIROOT}")
 set(MPI_LIBRARIES ${MPI_LIBRARY})
