+++ SourceFiles/Solver/pdsolve.cpp
@@ -15,6 +15,7 @@
 #include<algorithm>
 extern calMatrixOperations matoperat;
 extern pdGaussPt o_globGP;
+extern pdGaussPt o_sifGP;
 const double PI = acos(-1.);
 pdsolve::pdsolve(datModel & o_dat,int rank,int numProce)
 {
@@ -26,6 +27,7 @@ pdsolve::pdsolve(datModel & o_dat,int rank,int numProce)
 	cop_Ku = NULL; cop_F = NULL;
 	cip_ia = NULL, cip_ja = NULL, cdp_F = NULL, cdp_Ku = NULL, cdp_Ug = NULL, cdp_M = NULL;
 	cdp_FGlo = NULL, cdp_KuGlo = NULL;
+	cvp_inteNode = NULL;
 	//initial parameters for Newmark's method
 	cd_gamma = o_dat.cd_gamma;
 	cd_beta = o_dat.cd_beta;
@@ -96,6 +98,17 @@ pdsolve::~pdsolve()
 {
 	delete cop_D;
 	cop_D = NULL;
+
+	for (auto iter = cpv_pdbe.begin(); iter != cpv_pdbe.end(); ++iter)
+	{
+		if (*iter != nullptr)
+		{
+			delete[](*iter);
+			(*iter) = nullptr;
+		}
+	}
+	delete[] cvp_inteNode;
+	cvp_inteNode = NULL;
 }
 
 void pdsolve::setDatModel(datModel& o_dat)
@@ -108,42 +121,28 @@ void pdsolve::setDatModel(datModel& o_dat)
 	}
 
 	// below is order dependent;
-	setFEID_PDEID(o_dat);
-	Setdof_Index(o_dat);
-	setPDNODEandnumFami(o_dat);
-	if (o_dat.getTotnumFami()!=0)// if not pure FEM
-	{
-		findDomainDimen(o_dat);
-	/*	if (ci_rank==0)
-		{
-			printf("finished findDomainDimen......\n");
-		}*/
-		calVolumeOfNode(o_dat);//mpi
-		/*if (ci_rank == 0)
-		{
-			printf("finished calVolumeOfNode......\n");
-		}*/
-		setDeltaMaxMin(o_dat);//mpi
-		/*if (ci_rank == 0)
-		{
-			printf("finished setDeltaMaxMin......\n");
-		}*/
-		setBlockAndFami(o_dat);//mpi
-		/*if (ci_rank == 0)
-		{
-			printf("finished setBlockAndFami......\n");
-		}*/
-		initialBondState(o_dat);
-		/*if (ci_rank == 0)
-		{
-			printf("finished initialBondState......\n");
-		}*/
-		setNoFailRegion(o_dat);
-		/*if (ci_rank == 0)
-		{
-			printf("finished setNoFailRegion......\n");
-		}*/
-	}
+	Setdof_Index( o_dat);
+	setDeltaMaxMin_EleV_NV( o_dat);// find out the max and min Delta. Got ele volume and  node volume
+	findDomainDimen( o_dat);// get the domain dimension size;
+	setBlock( o_dat);// allocate memory for block, and put elements and elements into block;
+	initializePDE_FE( o_dat);//judge FE and PDE, then put the idx into vector, based on intersection between element and crack;
+	updatePDBE( o_dat);//update PDBE and assign the pointers, based on the civ_pdeIDX;
+	UpdateNodeType( o_dat);
[…]
