+++ HeaderFiles/Solver/fioFile.h
@@ -38,8 +38,8 @@ public:
 	void writeReslutsTOTAL_vtk_ASCII(datModel& o_dat, string nTstep);
 	void writeReslutsTOTAL_vtk_Binary(datModel& o_dat, string nTstep);
 private:
-	void writeUofNode(datModel &o_dat);
-	void writeSigofNode(datModel &o_dat);
+	void writeUofNode(datModel &o_dat, string nTstep);
+	void writeSigofNode(datModel &o_dat, string nTstep);
 	void writeResultsPD_vtk(datModel &o_dat);
 	void writeResultsFEM_vtk(datModel& o_dat);
 	void setTitle(datModel &o_dat);

[…]
