+++ examples/boxHole/boxHolecmd_static.in
@@ -4,12 +4,9 @@ MSHFILE boxHole.lis
 SOLVER STATIC
 # set solving for dynamic solver, for static solver, these cmd will be ignored
 # the order is: Keyword, time incremental, total time step, write results frequency, horizon factor, non-local factor;
-SETSOLVING 1 72  2  4  0.25
+SETSOLVING 1 72  2  6  0.25
 # Newmark's method: ON or OFF; for dynamic solver; default is OFF;
 NEWMARK OFF
-#FENSF ON/OFF; default is OFF, FE node serve as PD family members or not.
-# For dynamic solving, we suggest to set it ON; for quasi/-static solving, it may only affect solving time cost
-FENSF OFF
 # lumped mass: ON or OFF; Only available for explicit dynamic solver; default is OFF:
 #LUMPMASS ON
 # reset essential boundary condition;
