### 2021-12-16 #3668 @prashjha
Hi @QibangLiu, could you address these points:

1. Draft corrections (suggestions):

- Line 7: I will suggest removing single quote around PDLSM-FEM solver
- Line 8: "and 3D using MPI parallelism" instead of "3D by MPI technique"?
- Line 8, 9: "This cross-platform solver is written in a C++ language and includes ..."?
- Line 27: "that of a PD with a spherical neighborhood"?
- Line 27: "However, their PD ..."?
- Line 37: you have used "much" multiple times and further "much" sounds not so scientific. How about "... mesh relatively fast and utilizes less memory compared to pure PD model implementations."?
- Line 39: In "In many published PD...", please specify references as "many" does not sound right.
- Line 44: "... in an easy way similar to FEM without introducing fictitious layer"?
- Line 44: "... solver enables two criteria ..."?
- Line 45: In "for 2D crack ...", is it just for 2D or is it is true also for 3D?
- Line 49: "It can also be ..."?

2. Once you have updated the draft, can you also do a 'tagged' release of your code, and

3. archive the release using zenedo or other methods.

Once you are done, I will run a few commands and hand your paper to EiC for the final decision.
