+++ JOSS_paper/paper.md
@@ -36,7 +36,7 @@ PDLSM-FEM solver is user friendly and can solve the structures with uniform or n
 PDLSM-FEM is developed to perform fracture analysis of structures, which takes advantage of both PD and FEM. It can also be used to perform conventional finite element analysis, when the PD region of the model is shrunk to zero, which means there is no PD element.
 \autoref{fig:diag} illustrates the crack propagation simulation of a diagonal plate with an inclined pre-existing crack by the PDLSM-FEM solver. As it shows, PDLSM-FEM solver captures the crack growth path which has a good agreement with the experimental observation [@ayatollahi_analysis_2009]. The geometry and material properties of this example can be found in [@liu_simulating_2021].
 
-![Simulation and experimental results of a diagonal plate under displacement loading: displacement $U_y$ with deformed shape (left-top), stress $\sigma_y$ with deformed shape (right-top), damage $\varphi$ with underformed shape (left-bottom), and experimental crack path (right-bottom). \label{fig:diag}](diagExample.png){ width=90% }
+![Simulation and experimental results of a diagonal plate: displacement $u_y$ with deformed shape (left-top), stress $\sigma_y$ with deformed shape (right-top), damage $\varphi$ with underformed shape (left-bottom), and experimental crack path (right-bottom). \label{fig:diag}](diagExample.png){ width=90% }
 
 
 # References
